pyproject.toml
[project]
name = "bskyshell"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "atproto>=0.0.55",
    "getch>=1.0",
    "python-dotenv>=1.0.1",
    "rich>=13.9.4",
    "typer>=0.13.1",
]

[project.scripts]
bsky = "src.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[dependency-groups]
dev = [
    "mypy>=1.13.0",
    "pre-commit>=4.0.1",
    "ruff>=0.8.0",
]
